introduces pundit

+++ app/controllers/api/v1/playlist_items_controller.rb
@@ -1,25 +1,21 @@
 class Api::V1::PlaylistItemsController < ApplicationController
-  before_action :authenticate_user
-  before_action :authorize_user, only: %i[update destroy]
   def create
     #clean up logic for creating playlistItems and songs
-    songs = playlist_items_params[:songs].map { |item| Song.create!(item) }
-    playlist = Playlist.find(params[:playlist_id])
-    items =
-      songs.map { |item| PlaylistItem.create!(song: item, playlist: playlist) }
+    playlist_items_params[:songs].each do |item|
+      PlaylistItem.create!(song: Song.find_or_create_by_uri(item), playlist: playlist)
+    end
 
     render json: PlaylistSerializer.new(playlist).serializable_hash.to_json,
            status: :created
   end
 
   def update
-    #fix update logic to so that we can reorder items
+    authorize playlist_item
     if playlist_item.insert_at(position)
-      render json:
-               PlaylistItemSerializer
-                 .new(playlist_item)
-                 .serializable_hash
-                 .to_json,
+      render json: PlaylistItemSerializer
+               .new(playlist_item)
+               .serializable_hash
+               .to_json,
              status: :ok
     else
       render status: :unprocessable_entity
@@ -27,11 +23,12 @@ class Api::V1::PlaylistItemsController < ApplicationController
   end
 
   def destroy
+    authorize playlist_item
     if playlist_item.destroy
       head :no_content
     else
       render json: {
-               error: playlist_item.errors.messages
+               error: playlist_item.errors.messages,
              },
              status: :unprocessable_entity
     end
@@ -55,5 +52,9 @@ class Api::V1::PlaylistItemsController < ApplicationController
 
   def position
     playlist_items_params[:position].to_i
+  end
+
+  def playlist
+    @playlist ||= Playlist.find(params[:playlist_id])
   end
 end
